third_party/blink/web_tests/platform/win/external/wpt/offscreen-canvas/text/2d.text.measure.baselines-expected.txt

This is a testharness.js-based test.
FAIL Testing baselines for OffscreenCanvas assert_equals: ctx.measureText('A').getBaselines().ideographic === -39 (got -12[number], expected -39[number]) expected -39 but got -12
Harness: the test ran to completion.
